# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/watchlists/{watchlistId}/items":{"post":{"tags":["watchlist-controller"],"operationId":"addItem","parameters":[{"name":"watchlistId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddWatchlistItemRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WatchlistItemResponse"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/users/{userId}/watchlists":{"get":{"tags":["watchlist-controller"],"operationId":"getUserWatchlists","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WatchlistResponse"}}}}}}},"post":{"tags":["watchlist-controller"],"operationId":"createWatchlist","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWatchlistRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WatchlistResponse"}}}}}}},"/api/users/{userId}/library":{"get":{"tags":["user-movie-controller"],"operationId":"getAll","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserMovieResponse"}}}}}}},"post":{"tags":["user-movie-controller"],"operationId":"add","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddUserMovieRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserMovieResponse"}}}}}}},"/api/movies/import/{externalId}":{"post":{"tags":["movie-controller"],"operationId":"importMovie","parameters":[{"name":"externalId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieResponse"}}}}}}},"/api/watchlists/{watchlistId}/items/reorder":{"patch":{"tags":["watchlist-controller"],"operationId":"reorder","parameters":[{"name":"watchlistId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReorderWatchlistRequest"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/api/users/{userId}/library/{userMovieId}/status":{"patch":{"tags":["user-movie-controller"],"operationId":"updateStatus","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userMovieId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserMovieResponse"}}}}}}},"/api/users/{userId}/library/{userMovieId}/rating":{"patch":{"tags":["user-movie-controller"],"operationId":"updateRating","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userMovieId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRatingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserMovieResponse"}}}}}}},"/api/users/{userId}/library/{userMovieId}/liked":{"patch":{"tags":["user-movie-controller"],"operationId":"updateLiked","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userMovieId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLikedRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserMovieResponse"}}}}}}},"/api/watchlists/{watchlistId}":{"get":{"tags":["watchlist-controller"],"operationId":"getWatchlist","parameters":[{"name":"watchlistId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WatchlistResponse"}}}}}},"delete":{"tags":["watchlist-controller"],"operationId":"deleteWatchlist","parameters":[{"name":"watchlistId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/movies/{id}":{"get":{"tags":["movie-controller"],"operationId":"getMovie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieResponse"}}}}}}},"/api/movies/search":{"get":{"tags":["movie-controller"],"operationId":"searchMovies","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/MovieSearchRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieSearchItem"}}}}}}}},"/api/watchlists/{watchlistId}/items/{itemId}":{"delete":{"tags":["watchlist-controller"],"operationId":"removeItem","parameters":[{"name":"watchlistId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"itemId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/users/{userId}/library/{userMovieId}":{"delete":{"tags":["user-movie-controller"],"operationId":"delete","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"userMovieId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"AddWatchlistItemRequest":{"type":"object","properties":{"externalId":{"type":"string"}},"required":["externalId"]},"MovieResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"externalId":{"type":"string"},"title":{"type":"string"},"year":{"type":"integer","format":"int32"},"runtimeMinutes":{"type":"integer","format":"int32"},"posterUrl":{"type":"string"},"overview":{"type":"string"}}},"WatchlistItemResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"position":{"type":"integer","format":"int32"},"addedAt":{"type":"string","format":"date-time"},"movie":{"$ref":"#/components/schemas/MovieResponse"}}},"CreateUserRequest":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["email","password","username"]},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"username":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"CreateWatchlistRequest":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title"]},"WatchlistResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"items":{"type":"array","items":{"$ref":"#/components/schemas/WatchlistItemResponse"}}}},"AddUserMovieRequest":{"type":"object","properties":{"externalId":{"type":"string"},"status":{"type":"string","enum":["PLANNED","WATCHED"]}},"required":["externalId","status"]},"UserMovieResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"movie":{"$ref":"#/components/schemas/MovieResponse"},"status":{"type":"string","enum":["PLANNED","WATCHED"]},"rating":{"type":"integer","format":"int32"},"liked":{"type":"boolean"},"watchedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ReorderWatchlistRequest":{"type":"object","properties":{"orderedItemIds":{"type":"array","items":{"type":"integer","format":"int64"}}},"required":["orderedItemIds"]},"UpdateStatusRequest":{"type":"object","properties":{"status":{"type":"string","enum":["PLANNED","WATCHED"]}},"required":["status"]},"UpdateRatingRequest":{"type":"object","properties":{"rating":{"type":"integer","format":"int32","maximum":10,"minimum":0}},"required":["rating"]},"UpdateLikedRequest":{"type":"object","properties":{"liked":{"type":"boolean"}},"required":["liked"]},"MovieSearchRequest":{"type":"object","properties":{"query":{"type":"string"},"year":{"type":"integer","format":"int32"},"page":{"type":"integer","format":"int32","minimum":1},"size":{"type":"integer","format":"int32","minimum":1}},"required":["query"]},"MovieSearchItem":{"type":"object","properties":{"externalId":{"type":"string"},"title":{"type":"string"},"year":{"type":"integer","format":"int32"},"posterUrl":{"type":"string"},"overview":{"type":"string"}}}}}}