# src/main/java/ru/mawshu/movietracker/dto/CreateWatchlistRequest.java
package ru.mawshu.movietracker.dto;

import jakarta.validation.constraints.NotBlank;

public class CreateWatchlistRequest {
    @NotBlank
    private String title;

    private String description;

    public String getTitle() { return title; }
    public void setTitle(String title) { this.title = title; }

    public String getDescription() { return description; }
    public void setDescription(String description) { this.description = description; }
}