# src/main/java/ru/mawshu/movietracker/dto/UpdateLikedRequest.java
package ru.mawshu.movietracker.dto;

import jakarta.validation.constraints.NotNull;

public class UpdateLikedRequest {
    @NotNull
    private Boolean liked;

    public Boolean getLiked() { return liked; }
    public void setLiked(Boolean liked) { this.liked = liked; }
}